Extract GrpcRunner class
